# [chore] [exporter/splunkhec] Remove redundant state data and allocation