--- a/exporter/splunkhecexporter/client.go
+++ b/exporter/splunkhecexporter/client.go
@@ -130,8 +130,8 @@ func (c *client) pushLogDataInBatches(ctx context.Context, ld plog.Logs, headers
 		profilingLocalHeaders[k] = v
 	}
 
-	var bufState = makeBlankBufferState(c.config.MaxContentLengthLogs, !c.config.DisableCompression, c.config.MaxEventSize)
-	var profilingBufState = makeBlankBufferState(c.config.MaxContentLengthLogs, !c.config.DisableCompression, c.config.MaxEventSize)
+	var bufState *bufferState
+	var profilingBufState *bufferState
 	var permanentErrors []error
 
 	var rls = ld.ResourceLogs()
@@ -143,6 +143,9 @@ func (c *client) pushLogDataInBatches(ctx context.Context, ld plog.Logs, headers
 			var newPermanentErrors []error
 
 			if isProfilingData(ills.At(j)) {
+				if profilingBufState == nil {
+					profilingBufState = makeBlankBufferState(c.config.MaxContentLengthLogs, !c.config.DisableCompression, c.config.MaxEventSize)
+				}
 				if !c.config.ProfilingDataEnabled {
 					droppedProfilingDataRecords += ills.At(j).LogRecords().Len()
 					continue
@@ -150,6 +153,9 @@ func (c *client) pushLogDataInBatches(ctx context.Context, ld plog.Logs, headers
 				profilingBufState.resource, profilingBufState.library = i, j
 				newPermanentErrors, err = c.pushLogRecords(ctx, rls, profilingBufState, profilingLocalHeaders)
 			} else {
+				if bufState == nil {
+					bufState = makeBlankBufferState(c.config.MaxContentLengthLogs, !c.config.DisableCompression, c.config.MaxEventSize)
+				}
 				if !c.config.LogDataEnabled {
 					droppedLogRecords += ills.At(j).LogRecords().Len()
 					continue
@@ -159,7 +165,7 @@ func (c *client) pushLogDataInBatches(ctx context.Context, ld plog.Logs, headers
 			}
 
 			if err != nil {
-				return consumererror.NewLogs(err, c.subLogs(ld, bufState.bufFront, profilingBufState.bufFront))
+				return consumererror.NewLogs(err, c.subLogs(ld, bufState, profilingBufState))
 			}
 
 			permanentErrors = append(permanentErrors, newPermanentErrors...)
@@ -174,17 +180,17 @@ func (c *client) pushLogDataInBatches(ctx context.Context, ld plog.Logs, headers
 	}
 
 	// There's some leftover unsent non-profiling data
-	if bufState.containsData {
+	if bufState != nil && bufState.containsData {
 		if err := c.postEvents(ctx, bufState, headers); err != nil {
-			return consumererror.NewLogs(err, c.subLogs(ld, bufState.bufFront, profilingBufState.bufFront))
+			return consumererror.NewLogs(err, c.subLogs(ld, bufState, profilingBufState))
 		}
 	}
 
 	// There's some leftover unsent profiling data
-	if profilingBufState.containsData {
+	if profilingBufState != nil && profilingBufState.containsData {
 		if err := c.postEvents(ctx, profilingBufState, profilingLocalHeaders); err != nil {
 			// Non-profiling bufFront is set to nil because all non-profiling data was flushed successfully above.
-			return consumererror.NewLogs(err, c.subLogs(ld, nil, profilingBufState.bufFront))
+			return consumererror.NewLogs(err, c.subLogs(ld, nil, profilingBufState))
 		}
 	}
 
@@ -195,10 +201,8 @@ func (c *client) pushLogRecords(ctx context.Context, lds plog.ResourceLogsSlice,
 	res := lds.At(state.resource)
 	logs := res.ScopeLogs().At(state.library).LogRecords()
 
+	state.record = 0
 	for k := 0; k < logs.Len(); k++ {
-		if state.bufFront == nil {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
-		}
 		var b []byte
 
 		if c.config.ExportRaw {
@@ -237,11 +241,11 @@ func (c *client) pushLogRecords(ctx context.Context, lds plog.ResourceLogsSlice,
 		// Writing truncated bytes back to buffer.
 		accepted, e = state.accept(b)
 		if accepted {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
+			state.record = k
 			continue
 		}
 
-		state.bufFront = nil
+		state.record = k + 1
 		if e != nil {
 			permanentErrors = append(permanentErrors, consumererror.NewPermanent(
 				fmt.Errorf("error writing the event: %w", e)))
@@ -258,11 +262,8 @@ func (c *client) pushMetricsRecords(ctx context.Context, mds pmetric.ResourceMet
 	res := mds.At(state.resource)
 	metrics := res.ScopeMetrics().At(state.library).Metrics()
 
+	state.record = 0
 	for k := 0; k < metrics.Len(); k++ {
-		if state.bufFront == nil {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
-		}
-
 		// Parsing metric record to Splunk event.
 		events := mapMetricToSplunkEvent(res.Resource(), metrics.At(k), c.config, c.logger)
 		buf := bytes.NewBuffer(make([]byte, 0, c.config.MaxContentLengthMetrics))
@@ -306,11 +307,11 @@ func (c *client) pushMetricsRecords(ctx context.Context, mds pmetric.ResourceMet
 		// Writing truncated bytes back to buffer.
 		accepted, e = state.accept(b)
 		if accepted {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
+			state.record = k
 			continue
 		}
 
-		state.bufFront = nil
+		state.record = k + 1
 		if e != nil {
 			permanentErrors = append(permanentErrors, consumererror.NewPermanent(
 				fmt.Errorf("error writing the event: %w", e)))
@@ -327,11 +328,8 @@ func (c *client) pushTracesData(ctx context.Context, tds ptrace.ResourceSpansSli
 	res := tds.At(state.resource)
 	spans := res.ScopeSpans().At(state.library).Spans()
 
+	state.record = 0
 	for k := 0; k < spans.Len(); k++ {
-		if state.bufFront == nil {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
-		}
-
 		// Parsing span record to Splunk event.
 		event := mapSpanToSplunkEvent(res.Resource(), spans.At(k), c.config)
 		// JSON encoding event and writing to buffer.
@@ -362,11 +360,11 @@ func (c *client) pushTracesData(ctx context.Context, tds ptrace.ResourceSpansSli
 		// Writing truncated bytes back to buffer.
 		accepted, e = state.accept(b)
 		if accepted {
-			state.bufFront = &index{resource: state.resource, library: state.library, record: k}
+			state.record = k
 			continue
 		}
 
-		state.bufFront = nil
+		state.record = k + 1
 		if e != nil {
 			permanentErrors = append(permanentErrors, consumererror.NewPermanent(
 				fmt.Errorf("error writing the event: %w", e)))
@@ -397,7 +395,7 @@ func (c *client) pushMetricsDataInBatches(ctx context.Context, md pmetric.Metric
 			newPermanentErrors, err = c.pushMetricsRecords(ctx, rms, bufState, headers)
 
 			if err != nil {
-				return consumererror.NewMetrics(err, subMetrics(md, bufState.bufFront))
+				return consumererror.NewMetrics(err, subMetrics(md, bufState))
 			}
 
 			permanentErrors = append(permanentErrors, newPermanentErrors...)
@@ -407,7 +405,7 @@ func (c *client) pushMetricsDataInBatches(ctx context.Context, md pmetric.Metric
 	// There's some leftover unsent metrics
 	if bufState.containsData {
 		if err := c.postEvents(ctx, bufState, headers); err != nil {
-			return consumererror.NewMetrics(err, subMetrics(md, bufState.bufFront))
+			return consumererror.NewMetrics(err, subMetrics(md, bufState))
 		}
 	}
 
@@ -432,7 +430,7 @@ func (c *client) pushTracesDataInBatches(ctx context.Context, td ptrace.Traces, 
 			newPermanentErrors, err = c.pushTracesData(ctx, rts, bufState, headers)
 
 			if err != nil {
-				return consumererror.NewTraces(err, subTraces(td, bufState.bufFront))
+				return consumererror.NewTraces(err, subTraces(td, bufState))
 			}
 
 			permanentErrors = append(permanentErrors, newPermanentErrors...)
@@ -442,7 +440,7 @@ func (c *client) pushTracesDataInBatches(ctx context.Context, td ptrace.Traces, 
 	// There's some leftover unsent traces
 	if bufState.containsData {
 		if err := c.postEvents(ctx, bufState, headers); err != nil {
-			return consumererror.NewTraces(err, subTraces(td, bufState.bufFront))
+			return consumererror.NewTraces(err, subTraces(td, bufState))
 		}
 	}
 
@@ -456,45 +454,26 @@ func (c *client) postEvents(ctx context.Context, bufState *bufferState, headers 
 	return c.hecWorker.send(ctx, bufState, headers)
 }
 
-// subLogs returns a subset of `ld` starting from `profilingBufFront` for profiling data
-// plus starting from `bufFront` for non-profiling data. Both can be nil, in which case they are ignored
-func (c *client) subLogs(ld plog.Logs, bufFront *index, profilingBufFront *index) plog.Logs {
+// subLogs returns a subset of `ld` starting from `profilingState` for profiling data
+// plus starting from `state` for non-profiling data.
+func (c *client) subLogs(ld plog.Logs, state *bufferState, profilingState *bufferState) plog.Logs {
 	subset := plog.NewLogs()
-	if c.config.LogDataEnabled {
-		subLogsByType(ld, bufFront, subset, false)
+	if c.config.LogDataEnabled && state != nil {
+		subLogsByType(ld, state, subset, false)
 	}
-	if c.config.ProfilingDataEnabled {
-		subLogsByType(ld, profilingBufFront, subset, true)
+	if c.config.ProfilingDataEnabled && profilingState != nil {
+		subLogsByType(ld, profilingState, subset, true)
 	}
 
 	return subset
 }
 
-// subMetrics returns a subset of `md`starting from `bufFront`. It can be nil, in which case it is ignored
-func subMetrics(md pmetric.Metrics, bufFront *index) pmetric.Metrics {
-	subset := pmetric.NewMetrics()
-	subMetricsByType(md, bufFront, subset)
-
-	return subset
-}
-
-// subTraces returns a subset of `td`starting from `bufFront`. It can be nil, in which case it is ignored
-func subTraces(td ptrace.Traces, bufFront *index) ptrace.Traces {
-	subset := ptrace.NewTraces()
-	subTracesByType(td, bufFront, subset)
-
-	return subset
-}
-
-func subLogsByType(src plog.Logs, from *index, dst plog.Logs, profiling bool) {
-	if from == nil {
-		return // All the data of this type was sent successfully
-	}
-
+// subLogs returns a subset of logs starting the state.
+func subLogsByType(src plog.Logs, state *bufferState, dst plog.Logs, profiling bool) {
 	resources := src.ResourceLogs()
 	resourcesSub := dst.ResourceLogs()
 
-	for i := from.resource; i < resources.Len(); i++ {
+	for i := state.resource; i < resources.Len(); i++ {
 		newSub := resourcesSub.AppendEmpty()
 		resources.At(i).Resource().CopyTo(newSub.Resource())
 
@@ -502,8 +481,8 @@ func subLogsByType(src plog.Logs, from *index, dst plog.Logs, profiling bool) {
 		librariesSub := newSub.ScopeLogs()
 
 		j := 0
-		if i == from.resource {
-			j = from.library
+		if i == state.resource {
+			j = state.library
 		}
 		for jSub := 0; j < libraries.Len(); j++ {
 			lib := libraries.At(j)
@@ -521,8 +500,8 @@ func subLogsByType(src plog.Logs, from *index, dst plog.Logs, profiling bool) {
 			jSub++
 
 			k := 0
-			if i == from.resource && j == from.library {
-				k = from.record
+			if i == state.resource && j == state.library {
+				k = state.record
 			}
 
 			for kSub := 0; k < logs.Len(); k++ { //revive:disable-line:var-naming
@@ -533,15 +512,13 @@ func subLogsByType(src plog.Logs, from *index, dst plog.Logs, profiling bool) {
 	}
 }
 
-func subMetricsByType(src pmetric.Metrics, from *index, dst pmetric.Metrics) {
-	if from == nil {
-		return // All the data of this type was sent successfully
-	}
-
+// subMetrics returns a subset of metrics starting from the state.
+func subMetrics(src pmetric.Metrics, state *bufferState) pmetric.Metrics {
+	dst := pmetric.NewMetrics()
 	resources := src.ResourceMetrics()
 	resourcesSub := dst.ResourceMetrics()
 
-	for i := from.resource; i < resources.Len(); i++ {
+	for i := state.resource; i < resources.Len(); i++ {
 		newSub := resourcesSub.AppendEmpty()
 		resources.At(i).Resource().CopyTo(newSub.Resource())
 
@@ -549,8 +526,8 @@ func subMetricsByType(src pmetric.Metrics, from *index, dst pmetric.Metrics) {
 		librariesSub := newSub.ScopeMetrics()
 
 		j := 0
-		if i == from.resource {
-			j = from.library
+		if i == state.resource {
+			j = state.library
 		}
 		for jSub := 0; j < libraries.Len(); j++ {
 			lib := libraries.At(j)
@@ -563,8 +540,8 @@ func subMetricsByType(src pmetric.Metrics, from *index, dst pmetric.Metrics) {
 			jSub++
 
 			k := 0
-			if i == from.resource && j == from.library {
-				k = from.record
+			if i == state.resource && j == state.library {
+				k = state.record
 			}
 
 			for kSub := 0; k < metrics.Len(); k++ { //revive:disable-line:var-naming
@@ -573,17 +550,16 @@ func subMetricsByType(src pmetric.Metrics, from *index, dst pmetric.Metrics) {
 			}
 		}
 	}
+
+	return dst
 }
 
-func subTracesByType(src ptrace.Traces, from *index, dst ptrace.Traces) {
-	if from == nil {
-		return // All the data of this type was sent successfully
-	}
-
+func subTraces(src ptrace.Traces, state *bufferState) ptrace.Traces {
+	dst := ptrace.NewTraces()
 	resources := src.ResourceSpans()
 	resourcesSub := dst.ResourceSpans()
 
-	for i := from.resource; i < resources.Len(); i++ {
+	for i := state.resource; i < resources.Len(); i++ {
 		newSub := resourcesSub.AppendEmpty()
 		resources.At(i).Resource().CopyTo(newSub.Resource())
 
@@ -591,10 +567,10 @@ func subTracesByType(src ptrace.Traces, from *index, dst ptrace.Traces) {
 		librariesSub := newSub.ScopeSpans()
 
 		j := 0
-		if i == from.resource {
-			j = from.library
+		if i == state.resource {
+			j = state.library
 		}
-		for jSub := 0; j < libraries.Len(); j++ {
+		for ; j < libraries.Len(); j++ {
 			lib := libraries.At(j)
 
 			newLibSub := librariesSub.AppendEmpty()
@@ -602,19 +578,19 @@ func subTracesByType(src ptrace.Traces, from *index, dst ptrace.Traces) {
 
 			traces := lib.Spans()
 			tracesSub := newLibSub.Spans()
-			jSub++
 
 			k := 0
-			if i == from.resource && j == from.library {
-				k = from.record
+			if i == state.resource && j == state.library {
+				k = state.record
 			}
 
-			for kSub := 0; k < traces.Len(); k++ { //revive:disable-line:var-naming
+			for ; k < traces.Len(); k++ {
 				traces.At(k).CopyTo(tracesSub.AppendEmpty())
-				kSub++
 			}
 		}
 	}
+
+	return dst
 }
 
 func (c *client) stop(context.Context) error {